requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
pypdf
faiss-cpu
openai
numpy
pydantic
requests
python-multipart